Move convert behavior to variables package
Aggregate conversion implementations into the SDK rather than inside
the Pack implementation code.
